% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{batchLL}
\alias{batchLL}
\title{batchLL function provides the batch marking log-likelihood}
\usage{
batchLL(phi, p, R, begin_g, end_g)
}
\arguments{
\item{phi}{The probability of surviving and remaining in the population between occasions t and t +1, given an individual was alive and in the population at occasion t. This must be a number between 0 and 1.}

\item{p}{The probability of capture at occasion t. This must be a number between 0 and 1.}

\item{R}{The number of individuals marked and released at sampling occasion g from batch group g; g = 1,2,...,G. This must be an integer.}

\item{begin_g}{The beginning of the occasion.}

\item{end_g}{The end of the occasion.}
}
\value{
fr returns the log sum of the Hidden Markov Model.
}
\description{
batchLL function provides the batch marking log-likelihood
}
