% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{batchLogit}
\alias{batchLogit}
\title{batchLogit function}
\usage{
batchLogit(x)
}
\arguments{
\item{x}{This is an input numerical value i.e double.}
}
\value{
Returns a number between 0 and 1.
}
\description{
'batchLogit' provides the number between 0 and 1.
}
