% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{batchUnmark2Viterbi}
\alias{batchUnmark2Viterbi}
\title{batchUnmark2Viterbi function provides a wrapper for the batchUnmarkViterbi to compute the popuation abundance}
\usage{
batchUnmark2Viterbi(
  par,
  data,
  Umax,
  nBins,
  choiceModel = c("model1", "model2", "model3", "model4")
)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{data}{A capture-recapture data matrix or data frame.}

\item{Umax}{The maximum number of the unmarked individuals in the population for capture on any occasion.}

\item{nBins}{The number of bin size into which the matrix will be divided.}

\item{choiceModel}{This chooses among different models and allows for model selection}
}
\value{
Negative Log-likelihood value of the likelihood function
}
\description{
batchUnmark2Viterbi function provides a wrapper for the batchUnmarkViterbi to compute the popuation abundance
}
