% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{dbinpois}
\alias{dbinpois}
\title{Convolution of Poisson and Binomial for Batch}
\usage{
dbinpois(z, n, par)
}
\arguments{
\item{z}{This is the vector of numerical values}

\item{n}{The \code{nrow} of capture-recapture data matrix or data frame}

\item{par}{This is the vector of parameter values: average from Poisson distribution and probability of success from Binomial distribution}
}
\value{
f This is the output of the convolution from the Binomial and Poisson distributions
}
\description{
This is the convolution of Poisson and Binomial distributions
}
\details{
The convolution of Poisson and Binomial distribution helps us to compute the number of individuals that have survived from t-1 to t in the combined model while simultaneously computing the number of individuals recruited into the population at occasion t.

The survival is modeled as Binomial distribution and the recruitment as the Poisson distirubiton
}
