% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{delta_g}
\alias{delta_g}
\title{initial probability function}
\usage{
delta_g(R)
}
\arguments{
\item{R}{The number of individuals marked and released at sampling occasion g from batch group g; g = 1,2,...,G. This must be an integer.}
}
\value{
A vector of initial value with 1 at the observed position
}
\description{
initial probability function
}
