% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBatchMarking-package.R
\docType{package}
\name{extBatchMarking-package}
\alias{extBatchMarking}
\alias{extBatchMarking-package}
\title{extBatchMarking: Extended Batch Marking Models}
\description{
A system for batch-marking data analysis to estimate survival probabilities, capture probabilities, and enumerate the population abundance for both marked and unmarked individuals. The estimation of only marked individuals can be achieved through the batchMarkOptim() function. Similarly, the combined marked and unmarked can be achieved through the batchMarkUnmarkOptim() function. The algorithm was also implemented for the hidden Markov model encapsulated in batchMarkUnmarkOptim() to estimate the abundance of both marked and unmarked individuals in the population. The package is based on the paper: "Hidden Markov Models for Extended Batch Data" of Cowen et al. (2017) \doi{10.1111/biom.12701}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Olobatuyi/extBatchMarking_cov}
  \item Report bugs at \url{https://github.com/Olobatuyi/extBatchMarking_cov/issues}
}

}
\author{
\strong{Maintainer}: Kehinde Olobatuyi \email{olobatuyikenny@uvic.ca} (\href{https://orcid.org/0000-0002-4635-7895}{ORCID})

Authors:
\itemize{
  \item Simon Johns \email{simon737@outlook.com}
  \item Matthew RP Parker \email{mrparker909@gmail.com} (\href{https://orcid.org/0000-0003-3021-7959}{ORCID})
  \item Steve Hof \email{steven.Hof@icloud.com}
  \item Laura LE Cowen \email{lcowen@uvic.ca} (\href{https://orcid.org/0000-0002-0853-1450}{ORCID})
}

}
\keyword{internal}
