% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_gt}
\alias{gamma_gt}
\title{Transition State Probability
'gamma_gt' computes the transition probability matrix}
\arguments{
\item{R}{integer number of marked individuals released per occasion}

\item{phi}{double number. Survival probability of individuals}

\item{cores}{The number of cores on your machine.}
}
\value{
pR Returns the transition matrix
}
\description{
Transition State Probability
'gamma_gt' computes the transition probability matrix
}
