% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMarkfun.R
\name{plot.batchMarkOptim}
\alias{plot.batchMarkOptim}
\title{Plot Method for batchMarkUnmarkOptim Objects}
\usage{
\method{plot}{batchMarkOptim}(x, ...)
}
\arguments{
\item{x}{An object of class "Employee".}

\item{...}{Additional arguments passed to the print method.}
}
\description{
This function defines how objects of class "Employee" are printed.
}
