% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMark.R
\name{probs}
\alias{probs}
\title{State-dependent probability function}
\usage{
probs(r, p, R)
}
\arguments{
\item{r}{The number of individuals from batch group "g" recaptured at recapture occasion t; g = 1,2,...,G, t = g+1,...,T. This must be an integer.}

\item{p}{The probability of capture at occasion t. This must be a number between 0 and 1.}

\item{R}{The number of individuals marked and released at sampling occasion g from batch group g; g = 1,2,...,G. This must be an integer.}
}
\value{
PR diagonal matrix of the state-dependent probability.
}
\description{
'probs' computes the state-dependent transition matrix
}
