% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherLoach.R
\docType{data}
\name{WeatherLoach}
\alias{WeatherLoach}
\title{Weather Loach data}
\format{
\subsection{'Weather_loach'}{

A data frame with 10 rows indicating number of captures and 11 columns indicating recaptures
\describe{
\item{Weather Loach Data}{}
}
}
}
\usage{
WeatherLoach
}
\description{
Data from marked individuals captured on multiple occasions
The weather-loach study was described in detail by Huggin ()
Different colored batch tags were given to a random sample of unmarked individuals at each occasion.
}
\keyword{datasets}
